+++ src/main/java/org/synyx/urlaubsverwaltung/person/basedata/PersonBasedataRepository.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.person.masterdata;
+package org.synyx.urlaubsverwaltung.person.basedata;
 
 import org.springframework.data.jpa.repository.JpaRepository;
