+++ src/test/java/org/synyx/urlaubsverwaltung/application/statistics/ApplicationForLeaveStatisticsServiceTest.java
@@ -10,8 +10,8 @@ import org.synyx.urlaubsverwaltung.application.vacationtype.VacationTypeService;
 import org.synyx.urlaubsverwaltung.department.DepartmentService;
 import org.synyx.urlaubsverwaltung.person.Person;
 import org.synyx.urlaubsverwaltung.person.PersonService;
-import org.synyx.urlaubsverwaltung.person.masterdata.PersonBasedata;
-import org.synyx.urlaubsverwaltung.person.masterdata.PersonBasedataService;
+import org.synyx.urlaubsverwaltung.person.basedata.PersonBasedata;
+import org.synyx.urlaubsverwaltung.person.basedata.PersonBasedataService;
 import org.synyx.urlaubsverwaltung.web.FilterPeriod;
 
 import java.time.LocalDate;
